feat: improve file validation performance
