+++ deepset_cloud_sdk/workflows/sync_client/files.py
@@ -65,6 +65,7 @@ def upload(
     blocking: bool = True,
     timeout_s: int = 300,
     show_progress: bool = True,
+    recursive: bool = False,
 ) -> None:
     """Upload a folder to deepset Cloud.
 
@@ -85,6 +86,7 @@ def upload(
             blocking=blocking,
             timeout_s=timeout_s,
             show_progress=show_progress,
+            recursive=recursive,
         )
     )
