+++ pyproject.toml
@@ -33,6 +33,7 @@ dependencies = [
   "aiofiles==23.1.0",
   "tabulate==0.9.0",
   "tqdm==4.65.0",
+  "yaspin==2.3.0",
 ]
 
 [project.urls]
