+++ tests/data/upload_folder_nested/meta/example.txt.meta.json
@@ -1,0 +1,1 @@
+{"a": "b"}
